refactor(services): move logMessage to messages.service
