+++ src/app/services/card.services/trap-card.service.ts
@@ -1,7 +1,6 @@
 import { Player } from "../../models/player.model";
 import { playerLosesMoney } from "./shared-card.service";
-import { Messages } from "../messasges.service";
-import { logMessage } from "../info-board.service";
+import { Messages, logMessage } from "../messasges.service";
 import { TRAP_TAX } from "../../constants/prices";
 import { PLAYER_IS_ATTACKED_STEPS, PLAYER_IS_CHASED_STEPS } from "../../constants/game-constants";
